smiv: Add HW support for decompression of CSR weights.
This is implemented as a separate accelerator block that runs just prior
to starting an FC layer. The CSR decompression block reads in the packed
inputs and indices into one of the scratchpads, zeros out the part of
the UMEM it will need to modify, and writes the decompressed data into
the UMEM. Then, the FC block can start to run without needing to load
the weights.

+++ nnet_lib/src/arch/smiv/convolution.c
@@ -88,7 +88,7 @@ void reduction_hw(float* spad0,
         }
     }
     printf("reduction results:\n");
-    for (int i = 0; i < result_bytes / sizeof(float); i++) {
+    for (int i = 0; i < result_bytes / (int)sizeof(float); i++) {
         printf("%f, ", host_result[i]);
     }
     printf("\n");
